requirements.txt
torch
torchvision
numpy
Pillow
torchmetrics